+++ lib/l10n/intl_en.arb
@@ -178,6 +178,7 @@
   "depositProcessStep2": "Deposit in progress",
   "depositProcessStep3": "Process finished",
   "farmDepositControlAmountEmpty": "Please enter the amount",
+  "farmDepositControlLPTokenAmountExceedBalance": "The amount entered exceeds your balance.",
   "farmDepositConfirmTitle": "Confirmation",
   "farmDepositFormTitle": "Deposit",
   "farmDepositProcessInterruptionWarning": "Are you sure you want to stop the process?",
@@ -197,6 +198,7 @@
   "farmWithdrawSuccessInfo": "Withdraw has been done successfully",
   "farmWithdrawInProgressConfirmAEWallet": "Please, confirm a transaction in your Archethic Wallet",
   "farmWithdrawProcessInProgress": "Please wait.",
+  "farmWithdrawControlLPTokenAmountExceedDeposited": "The amount entered exceeds the deposited amount.",
   "btn_farm_withdraw": "Withdraw",
   "btn_confirm_farm_withdraw": "Confirm",
   "claimProcessStep0": "Process not started",
