+++ lib/ui/views/liquidity_add/layouts/components/liquidity_add_textfield_token_2_amount.dart
@@ -1,6 +1,9 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
 import 'package:aedex/ui/themes/dex_theme_base.dart';
 import 'package:aedex/ui/views/liquidity_add/bloc/provider.dart';
+import 'package:aedex/ui/views/util/components/dex_btn_half.dart';
+import 'package:aedex/ui/views/util/components/dex_btn_max.dart';
+import 'package:aedex/ui/views/util/components/dex_token_balance.dart';
 import 'package:aedex/ui/views/util/generic/formatters.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -64,71 +67,114 @@ class _LiquidityAddToken2AmountState
       _updateAmountTextController();
     }
 
-    return SizedBox(
-      width: DexThemeBase.sizeBoxComponentWidth,
-      child: Row(
-        children: [
-          Expanded(
-            child: DecoratedBox(
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(20),
-              ),
-              child: Row(
-                children: [
-                  Expanded(
-                    child: DecoratedBox(
-                      decoration: BoxDecoration(
-                        borderRadius: BorderRadius.circular(
-                          10,
-                        ),
-                        border: Border.all(
-                          color: Theme.of(context).colorScheme.primaryContainer,
-                          width: 0.5,
-                        ),
-                        gradient: LinearGradient(
-                          colors: [
-                            Theme.of(context)
-                                .colorScheme
-                                .background
-                                .withOpacity(1),
-                            Theme.of(context)
-                                .colorScheme
-                                .background
-                                .withOpacity(0.3),
-                          ],
-                          stops: const [0, 1],
-                        ),
-                      ),
-                      child: TextField(
-                        style: Theme.of(context).textTheme.titleMedium,
-                        autocorrect: false,
-                        controller: tokenAmountController,
-                        onChanged: (text) async {
-                          await liquidityAddNotifier.setToken2Amount(
-                            double.tryParse(text.replaceAll(' ', '')) ?? 0,
-                          );
-                        },
-                        focusNode: tokenAmountFocusNode,
-                        textAlign: TextAlign.left,
-                        textInputAction: TextInputAction.next,
-                        keyboardType: TextInputType.text,
-                        inputFormatters: <TextInputFormatter>[
-                          AmountTextInputFormatter(precision: 8),
-                          LengthLimitingTextInputFormatter(10),
-                        ],
-                        decoration: const InputDecoration(
-                          border: InputBorder.none,
-                          contentPadding: EdgeInsets.only(left: 10),
-                        ),
-                      ),
-                    ),
+    return Column(
+      children: [
+        SizedBox(
+          width: DexThemeBase.sizeBoxComponentWidth,
+          child: Row(
+            children: [
+              Expanded(
+                child: DecoratedBox(
+                  decoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(20),
                   ),
-                ],
+                  child: Row(
+                    children: [
+                      Expanded(
+                        child: DecoratedBox(
+                          decoration: BoxDecoration(
+                            borderRadius: BorderRadius.circular(
+                              10,
+                            ),
+                            border: Border.all(
+                              color: Theme.of(context)
+                                  .colorScheme
+                                  .primaryContainer,
+                              width: 0.5,
+                            ),
+                            gradient: LinearGradient(
+                              colors: [
+                                Theme.of(context)
+                                    .colorScheme
+                                    .background
+                                    .withOpacity(1),
+                                Theme.of(context)
+                                    .colorScheme
+                                    .background
+                                    .withOpacity(0.3),
+                              ],
+                              stops: const [0, 1],
+                            ),
+                          ),
+                          child: TextField(
+                            style: Theme.of(context).textTheme.titleMedium,
+                            autocorrect: false,
+                            controller: tokenAmountController,
+                            onChanged: (text) async {
+                              await liquidityAddNotifier.setToken2Amount(
+                                double.tryParse(text.replaceAll(' ', '')) ?? 0,
+                              );
+                            },
+                            focusNode: tokenAmountFocusNode,
+                            textAlign: TextAlign.left,
+                            textInputAction: TextInputAction.next,
+                            keyboardType: TextInputType.text,
+                            inputFormatters: <TextInputFormatter>[
+                              AmountTextInputFormatter(precision: 8),
+                              LengthLimitingTextInputFormatter(10),
+                            ],
+                            decoration: const InputDecoration(
+                              border: InputBorder.none,
+                              contentPadding: EdgeInsets.only(left: 10),
+                            ),
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
               ),
-            ),
+            ],
           ),
-        ],
-      ),
+        ),
+        const SizedBox(
+          height: 5,
+        ),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            DexTokenBalance(
+              tokenBalance: liquidityAdd.token2Balance,
+              tokenSymbol: liquidityAdd.token2 == null
+                  ? ''
+                  : liquidityAdd.token2!.symbol,
+            ),
+            Row(
+              children: [
+                DexButtonHalf(
+                  balanceAmount: liquidityAdd.token2Balance,
+                  onTap: () => ref
+                      .read(
+                        LiquidityAddFormProvider.liquidityAddForm.notifier,
+                      )
+                      .setToken2AmountHalf(),
+                ),
+                const SizedBox(
+                  width: 10,
+                ),
+                DexButtonMax(
+                  balanceAmount: liquidityAdd.token2Balance,
+                  onTap: () => ref
+                      .read(
+                        LiquidityAddFormProvider.liquidityAddForm.notifier,
+                      )
+                      .setToken2AmountMax(ref),
+                ),
+              ],
+            ),
+          ],
+        ),
+      ],
     );
   }
 }
