+++ lib/ui/views/liquidity_remove/layouts/components/liquidity_remove_form_sheet.dart
@@ -5,11 +5,8 @@ import 'package:aedex/ui/views/liquidity_remove/layouts/components/liquidity_rem
 import 'package:aedex/ui/views/liquidity_remove/layouts/components/liquidity_remove_textfield_lp_token_amount.dart';
 import 'package:aedex/ui/views/pool_list/pool_list_sheet.dart';
 import 'package:aedex/ui/views/util/components/dex_btn_close.dart';
-import 'package:aedex/ui/views/util/components/dex_btn_half.dart';
-import 'package:aedex/ui/views/util/components/dex_btn_max.dart';
 import 'package:aedex/ui/views/util/components/dex_btn_validate.dart';
 import 'package:aedex/ui/views/util/components/dex_error_message.dart';
-import 'package:aedex/ui/views/util/components/dex_token_balance.dart';
 import 'package:aedex/ui/views/util/components/pool_info_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/localizations.dart';
@@ -77,45 +74,6 @@ class LiquidityRemoveFormSheet extends ConsumerWidget {
                         height: 20,
                       ),
                       const LiquidityRemoveLPTokenAmount(),
-                      const SizedBox(
-                        height: 5,
-                      ),
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          DexTokenBalance(
-                            tokenBalance: liquidityRemove.lpTokenBalance,
-                            tokenSymbol: liquidityRemove.lpToken == null
-                                ? ''
-                                : liquidityRemove.lpToken!.symbol,
-                          ),
-                          Row(
-                            children: [
-                              DexButtonHalf(
-                                balanceAmount: liquidityRemove.lpTokenBalance,
-                                onTap: () => ref
-                                    .read(
-                                      LiquidityRemoveFormProvider
-                                          .liquidityRemoveForm.notifier,
-                                    )
-                                    .setLpTokenAmountHalf(),
-                              ),
-                              const SizedBox(
-                                width: 10,
-                              ),
-                              DexButtonMax(
-                                balanceAmount: liquidityRemove.lpTokenBalance,
-                                onTap: () => ref
-                                    .read(
-                                      LiquidityRemoveFormProvider
-                                          .liquidityRemoveForm.notifier,
-                                    )
-                                    .setLpTokenAmountMax(),
-                              ),
-                            ],
-                          ),
-                        ],
-                      ),
                       const SizedBox(
                         height: 10,
                       ),
