+++ lib/ui/views/liquidity_remove/layouts/components/liquidity_remove_textfield_lp_token_amount.dart
@@ -1,5 +1,8 @@
 import 'package:aedex/ui/themes/dex_theme_base.dart';
 import 'package:aedex/ui/views/liquidity_remove/bloc/provider.dart';
+import 'package:aedex/ui/views/util/components/dex_btn_half.dart';
+import 'package:aedex/ui/views/util/components/dex_btn_max.dart';
+import 'package:aedex/ui/views/util/components/dex_token_balance.dart';
 import 'package:aedex/ui/views/util/generic/formatters.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -65,71 +68,122 @@ class _LiquidityRemoveLPTokenAmountState
       _updateAmountTextController();
     }
 
-    return SizedBox(
-      width: DexThemeBase.sizeBoxComponentWidth,
-      child: Row(
-        children: [
-          Expanded(
-            child: DecoratedBox(
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(20),
-              ),
-              child: Row(
-                children: [
-                  Expanded(
-                    child: DecoratedBox(
-                      decoration: BoxDecoration(
-                        borderRadius: BorderRadius.circular(
-                          10,
-                        ),
-                        border: Border.all(
-                          color: Theme.of(context).colorScheme.primaryContainer,
-                          width: 0.5,
-                        ),
-                        gradient: LinearGradient(
-                          colors: [
-                            Theme.of(context)
-                                .colorScheme
-                                .background
-                                .withOpacity(1),
-                            Theme.of(context)
-                                .colorScheme
-                                .background
-                                .withOpacity(0.3),
-                          ],
-                          stops: const [0, 1],
-                        ),
-                      ),
-                      child: TextField(
-                        style: Theme.of(context).textTheme.titleMedium,
-                        autocorrect: false,
-                        controller: tokenAmountController,
-                        onChanged: (text) async {
-                          await liquidityRemoveNotifier.setLPTokenAmount(
-                            double.tryParse(text.replaceAll(' ', '')) ?? 0,
-                          );
-                        },
-                        focusNode: tokenAmountFocusNode,
-                        textAlign: TextAlign.left,
-                        textInputAction: TextInputAction.next,
-                        keyboardType: TextInputType.text,
-                        inputFormatters: <TextInputFormatter>[
-                          AmountTextInputFormatter(precision: 8),
-                          LengthLimitingTextInputFormatter(10),
-                        ],
-                        decoration: const InputDecoration(
-                          border: InputBorder.none,
-                          contentPadding: EdgeInsets.only(left: 10),
-                        ),
-                      ),
-                    ),
+    return Column(
+      children: [
+        SizedBox(
+          width: DexThemeBase.sizeBoxComponentWidth,
+          child: Row(
+            children: [
+              Expanded(
+                child: DecoratedBox(
+                  decoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(20),
                   ),
-                ],
+                  child: Row(
+                    children: [
+                      Expanded(
+                        child: DecoratedBox(
+                          decoration: BoxDecoration(
+                            borderRadius: BorderRadius.circular(
+                              10,
+                            ),
+                            border: Border.all(
+                              color: Theme.of(context)
+                                  .colorScheme
+                                  .primaryContainer,
+                              width: 0.5,
+                            ),
+                            gradient: LinearGradient(
+                              colors: [
+                                Theme.of(context)
+                                    .colorScheme
+                                    .background
+                                    .withOpacity(1),
+                                Theme.of(context)
+                                    .colorScheme
+                                    .background
+                                    .withOpacity(0.3),
+                              ],
+                              stops: const [0, 1],
+                            ),
+                          ),
+                          child: TextField(
+                            style: Theme.of(context).textTheme.titleMedium,
+                            autocorrect: false,
+                            controller: tokenAmountController,
+                            onChanged: (text) async {
+                              await liquidityRemoveNotifier.setLPTokenAmount(
+                                double.tryParse(text.replaceAll(' ', '')) ?? 0,
+                              );
+                            },
+                            focusNode: tokenAmountFocusNode,
+                            textAlign: TextAlign.left,
+                            textInputAction: TextInputAction.next,
+                            keyboardType: TextInputType.text,
+                            inputFormatters: <TextInputFormatter>[
+                              AmountTextInputFormatter(precision: 8),
+                              LengthLimitingTextInputFormatter(10),
+                            ],
+                            decoration: const InputDecoration(
+                              border: InputBorder.none,
+                              contentPadding: EdgeInsets.only(left: 10),
+                            ),
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
               ),
-            ),
+            ],
           ),
-        ],
-      ),
+        ),
+        const SizedBox(
+          height: 5,
+        ),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            DexTokenBalance(
+              tokenBalance: liquidityRemove.lpTokenBalance,
+              tokenSymbol: liquidityRemove.lpToken == null
+                  ? ''
+                  : liquidityRemove.lpToken!.symbol,
+            ),
+            Row(
+              children: [
+                DexButtonHalf(
+                  balanceAmount: liquidityRemove.lpTokenBalance,
+                  onTap: () {
+                    ref
+                        .read(
+                          LiquidityRemoveFormProvider
+                              .liquidityRemoveForm.notifier,
+                        )
+                        .setLpTokenAmountHalf();
+                    _updateAmountTextController();
+                  },
+                ),
+                const SizedBox(
+                  width: 10,
+                ),
+                DexButtonMax(
+                  balanceAmount: liquidityRemove.lpTokenBalance,
+                  onTap: () {
+                    ref
+                        .read(
+                          LiquidityRemoveFormProvider
+                              .liquidityRemoveForm.notifier,
+                        )
+                        .setLpTokenAmountMax();
+                    _updateAmountTextController();
+                  },
+                ),
+              ],
+            ),
+          ],
+        ),
+      ],
     );
   }
 }
